Add French and Arabic language support
