+++ lightrag_webui/src/i18n.ts
@@ -4,34 +4,44 @@ import { useSettingsStore } from '@/stores/settings'
 
 import en from './locales/en.json'
 import zh from './locales/zh.json'
+import fr from './locales/fr.json'
+import ar from './locales/ar.json'
 
-// Function to sync i18n with store state
-export const initializeI18n = async (): Promise<typeof i18n> => {
-  // Get initial language from store
-  const initialLanguage = useSettingsStore.getState().language
+const getStoredLanguage = () => {
+  try {
+    const settingsString = localStorage.getItem('settings-storage')
+    if (settingsString) {
+      const settings = JSON.parse(settingsString)
+      return settings.state?.language || 'en'
+    }
+  } catch (e) {
+    console.error('Failed to get stored language:', e)
+  }
+  return 'en'
+}
 
-  // Initialize with store language
-  await i18n.use(initReactI18next).init({
+i18n
+  .use(initReactI18next)
+  .init({
     resources: {
       en: { translation: en },
-      zh: { translation: zh }
+      zh: { translation: zh },
+      fr: { translation: fr },
+      ar: { translation: ar }
     },
-    lng: initialLanguage,
+    lng: getStoredLanguage(), // 使用存储的语言设置
     fallbackLng: 'en',
     interpolation: {
       escapeValue: false
     }
   })
 
-  // Subscribe to language changes
-  useSettingsStore.subscribe((state) => {
-    const currentLanguage = state.language
-    if (i18n.language !== currentLanguage) {
-      i18n.changeLanguage(currentLanguage)
-    }
-  })
-
-  return i18n
-}
+// Subscribe to language changes
+useSettingsStore.subscribe((state) => {
+  const currentLanguage = state.language
+  if (i18n.language !== currentLanguage) {
+    i18n.changeLanguage(currentLanguage)
+  }
+})
 
 export default i18n
